Rearrange options and update docs :children_crossing: :memo:
Overrides are now separate from the wallpaper constructors, which
makes more sense to me because the override has no relation to the
wallpaper itself.

Used mkRenamedOptionModule so that our warnings follow the standard
format.

I moved the documentation for wallpaper constructors to the option
description as this is more convenient when scrolling through
the list of options. It should also show up if someone generates a
full NixOS manual while having Stylix installed.