Implement black box type :recycle:
This introduces the unpack function, which simplifies the pattern
of writing this:

if «is video»
then «do something for video»
else if «is animation»
then «do something for animation»
else «do something for image»

into this:

«wallpaper».unpack {
  video = { file, ... }: «do something with the file»;
  animation = { file, ... }: «do something with the file»;
  image = { file, ... }: «do something with the file»;
}

A benefit of this is that it prevents you from writing the
if chain backwards:

if «is image»
then «do something for image»
...
else «do something for video»

Which would break if we added a new type because it tries to
treat the new type a video.

This system also allows us to have multiple fallbacks, so for example
if a video is unsupported it could be converted to a short animation
rather than going straight to a still image.