doc: add mention of `lib.stylix.pixel` to tricks (#327)
The function implements a fairly popular feature and is pretty old, but never gets mentioned in the docs. In fact, I only found it by randomly going through the sources. As a result, a person I spoke to wrote the section above describing a re-implementation of the function, which does work but doesn't need to be the default approach.

This change will improve discoverability of the codebase and direct users to a more standardized implementation.

P.S. `base16-schemes` changed `catppuccin.yaml` to a couple of variant themes, which is represented in this commit.