emacs: explicitly set font size
The emacs module currently doesn't set the font size, only the colours and font
family. This PR sets it to `sizes.terminal` (taking a lead from the `guifont`
setting in the `vim` module) using `font-spec`, which is a more flexible way to
specify fonts in elisp. To indicate the size is points it needs to be a floating
point value.