Don't split autoloaded modules into a separate derivation
Apparently `../modules` is creating a separate derivation that
contains only that folder, so it's now separate from the flake source.
But this transient derivation isn't mentioned explicitly in the flake
outputs. It makes it impossible to target those modules in
`disabledModules` directive.

For example, after this change is applied, users can solve issues like
https://github.com/danth/stylix/issues/577 locally, by just adding
the following snippet to their configuration:

    disabledModules = [ "${inputs.stylix}/modules/regreet/nixos.nix" ];