qt: add temporary warnings for plasma6
This adds temporary warnings when plasma6 is enabled and will be
replaced with actual handling of plasma6 when styling solutions are
added upstream.
Currently, this requires a manual work around that is only reasonable on
home-manager and not on NixOS. Users will be deferred to find
workarounds in the issue when they encounter this warning.