stylix: fix palette generator caching
This was broken in efb734ff436d7ba9df6648d3c9f3ce93a4b0f971,
where the added indirection prevents the output of the palette
generator becoming part of the final system.

Using `highestPrio` as a condition should prevent the palette
generator being forced to run when the output is not used, which
is the issue that the indirection was meant to solve.