stylix: fix palette generator caching (#867)
Ensure the palette generator is properly cached by fixing the bug
introduced in commit efb734ff436d ("Prevent scheme generation from
running when result is not used (#56)"), where the added indirection
prevents the output of the palette generator becoming part of the final
system.

Using highestPrio as a condition should prevent the palette generator
being forced to run when the output is not used, which is the issue that
the indirection was meant to solve.

Link: https://github.com/danth/stylix/pull/867

Reviewed-by: NAHO <90870942+trueNAHO@users.noreply.github.com>
(cherry picked from commit 9343b660c914a95d4a90f6bd40cda7c87b401445)