stylix: add editorconfig
> EditorConfig helps maintain consistent coding styles for multiple
developers working on the same project across various editors and IDEs.

Based on hm and nixpkgs editorconfig. Not sure about how many spaces to
indent css and html/xml/svg files (it's not consistent in the codebase).
A future pr should probably standardize the mustache file extensions (so
all css files are `.css.mustache` for example).