ci: ignore Cachix errors (#952)
This commit allows CI jobs to degrade gracefully if Cachix has issues,
by skipping uploads and/or building packages from scratch rather than
failing the job.

This is in response to the recent failures on `master`, which appear to
be caused by a 502 (Bad Gateway) response while uploading some of the
build results.