doc: reduce overuse of alerts (#1041)
Using too many "info" and "warning" boxes makes them less effective when
they're needed to draw attention to something genuinely important.

This commit adds important boxes to modules which require an external
module to be imported, similar to the box already used on the "Firefox and
derivatives" page.

It changes the warning when a module has no maintainers to a regular
paragraph.

The note about `stylix.enable` being new is removed, since this has been
around for a while now.