stylix: test the global enable option using assertions
This adds a pair of assertions to each testbed: that the system is not
changed in any way when Stylix is disabled, and that it is changed when
Stylix is enabled.

I originally wanted this to compare the value of every option, but that
uncovered a variety of errors in upstream code which is not normally
evaluated - things like `meta.maintainers` definitions using undefined
variables. I realised it would also require filtering out certain
subtrees, such as the stylix.enable option itself, which we expect to
change.

In the end, I opted to just compare the toplevel derivations instead.