ci: limit to one evaluation worker
Testbeds need a lot of memory to evaluate, so following the number
of cores is not the best idea.

I originally tried to scale based on the amount of memory available,
but this is platform dependent so didn't work on MacOS. It's also not
the best idea since other processes (such as our build jobs) may
allocate memory after the value was read.

I think it makes the most sense to just use one evaluation worker;
`nix flake check` does this anyway but by using `nix-fast-build` we
are still gaining the ability to skip builds which are in the binary
cache. (Eventually, this may become a feature in Nix itself.)