stylix: mock `stylix.targets` options in `clean` testbed
Some testbeds need to configure options for their target, for example
the Firefox testbed which enables Firefox GNOME theme. This fails when
Stylix is not installed (which we use to determine the original state
of the system, to see if Stylix changed anything). We can't just
exclude the testbed configuration from the clean system, because then
it would always change due to the target software being added.