doc: ensure code blocks are safely fenced
Markdown uses ``` to start and end fenced code blocks. However
code blocks can themselves contain ``` sequences.

To deal with this, markdown allows code block fences to be of arbitrary
length; any sequence of fence-chars that is shorter than the opening
fence will not terminate the code block.

We can scan through the code block text to find the longest sequence of
`-chars to ensure the opening/closing fences are longer than the longest
in the code block.