stylix: use a cheaper module eval for the `enable` option
The enable option must be evaluated for things like `nix flake show`, as
it determines which testbeds are included in the docs.

We can improve performance by skipping all the NixOS and home-manager
warnings and assertions that a system configuration usually evaluates.

We can further improve performance by not even importing the NixOS or
home-manager modules at all! This has some downsides; the `enable`
option cannot read `config` for example, and overlays will not be
applied to `pkgs`.

But it is worth it for the performance improvement when evaluating the
flake outputs.