stylix: use a cheaper module eval for the `enable` option
The `stylix.testbed.enable` option is needed to determine which testbeds
to include in flake outputs. However evaluating a NixOS system is
expensive, and significantly impacts the performance of `nix flake show`.

We work around this issue by adding minimal configuration, which
evaluates the testbed module on its own, without any nixos modules.

This has plenty of downsides, but we're unlikely to run into any of them
here.

Potential future issues:

1. A testbed module requiring a non-standard module arg.
   This could be solved by adding stubs to `_module.args`.

2. The enable option depending on an overlayed `pkgs` instance.
   This could be solved by implementing the `nixpkgs.overlays` option.

3. The enable option depending on another Stylix or NisOS option.
   This cannot be supported without massive performance hits.