stylix: refactor `base16Scheme` and `lib.stylix.colors` assertion (#1446)
Instead of having a nullable `base16Scheme` option, the default will
throw when `stylix.image` is null. Not only does this better represent
the "required" nature of this option, it also ensures the assertion is
evaluated early enough to actually be checked.

This causes the assertion to evaluate when the default definition is
read, instead of later when the `assertions` option is checked.

Link: https://github.com/nix-community/stylix/pull/1446

Tested-by: https://github.com/NovaViper
Approved-by: https://github.com/NovaViper
Reviewed-by: NAHO <90870942+trueNAHO@users.noreply.github.com>